New preflight checks and bug fix
NEW PREFLIGHT: If 4.1.1g is detected, root user is required. rescue-user can't write to /tmp in 4.1.1g. This is fixed in 4.2 and later.
NEW PREFLIGHT: In addition to disk space check, check for large files in /logs/k8_infra/eventmonitoring specifically.
BUG FIX: nd_validation_debug.log was failing to get bundled in final results bundle when running the script in WSL environments.